Added Book details
This is cool because it is data binding a FormView inside a ListView
using the book id set by the ListView! ListView stores the book id in a
hidden field and a Control value provider is used in the Controller to
locate it.
Also, there's no extra database calls as the first Controller method for
the ListView means the books are all cached in the DbContext - making
sure the same Controller is used for a single request (Page caches the
Controller)