Start of Progressive Enhancement
Wrap different areas in UpdatePanels. Set UpdateMode to Conditional and
ChildrenAsTriggers to false so they're not automatically updated. Use
NavigationDataTriggers to determine when to update.
Note that title doesn't trigger update of SortAndFilterPanel because
it's going to auto search - if it updated panel then focus would be
lost.
Need to set ClientIDMode to Auto when set PostBack to true on
NavigationHyperLinks because of .NET bug with __doPostBack and Ajax.